fixed karol's bug report from 09042013

+++ Modules/PortalModule/recommendationmanager.cpp
@@ -100,7 +100,7 @@ IResponse::Status RecommendationManager::createRecomm(IRequest *req, QString cur
         //username is solved inside createActivity method
         ac.activity_type = Activity::RECOMMENDATION;
         ac.content = absolute_uri + ";" + title;
-        ac.group_id = group_id.toInt();
+        ac.group_id = group_id;
         ac.object_id = uid;
         ac.user_id = curUser_id;
